requirements.txt
fastapi~=0.121.1
uvicorn~=0.38.0
pydantic~=2.12.4
SQLAlchemy~=2.0.44
python-jose~=3.5.0
PyYAML~=6.0.2
passlib~=1.7.4
openai~=2.7.1
python-dotenv~=1.2.1
asyncpg~=0.30.0
